% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{add_pilot_titles}
\alias{add_pilot_titles}
\title{Add titles to a plot using the Cpilot theme style}
\usage{
add_pilot_titles(
  plot,
  title = NULL,
  subtitle = NULL,
  title_family = "Avenir Next Demi Bold",
  subtitle_family = "Avenir Next",
  title_size = 17,
  subtitle_size = 12,
  title_color = "#404040",
  subtitle_color = "#404040",
  background_color = "#ffffff"
)
}
\arguments{
\item{plot}{A ggplot2 plot object to which titles will be added.}

\item{title}{A string containing the title to add to the plot. Use NULL if
you do not want a title.}

\item{subtitle}{A string containing the subtitle to add to the plot. Use
NULL if you do not want a subtitle.}

\item{title_family}{A string indicating the font-family to use for the
title. The default is "National-LFSN Semibd".}

\item{subtitle_family}{A string indicating the font-family to use for the
subtitle. The default is "National-LFSN Book".}

\item{title_size}{An integer indicating the font size to use for the title
in points. The default is 18 points.}

\item{subtitle_size}{An integer indicating the font size to use for the
subtitle in points. The default is 14 points.}

\item{title_color}{An RGB hex string indicating the color to use for the
title. The default is "#006548".}

\item{subtitle_color}{An RGB hex string indicating the color to use for the
subtitle. The default is "#006548".}

\item{background_color}{An RGB hex string indicating the color to use for
the background. The default is "#f0eeed".}
}
\value{
A copy of the input plot with a title and/or subtitle added.
}
\description{
Use this function to add titles to a plot that uses \code{theme_pilot}.
Using this functin to set the title and/or subtitle will ensure that title
elements are aligned with the left-hand edge of the y-axis text, rather than
the left-hand edge of the plotting area.
}
\details{
To use the function, first create a plot using \code{theme_pilot}
without setting the title or subtitle, capturing the plot object in a
variable. Pass the plot to \code{add_pilot_titles} along with the title
and subtitle you want to use. This function will return the same plot with
titles added to the top and the correct spacing between the titles and the
rest of the plot area.

Arguments are also available to set the font properties of the title and
subtitle elements of the plot.
}
