% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{annotate_pilot}
\alias{annotate_pilot}
\title{A wrapper for the annotate function in ggplot2}
\usage{
annotate_pilot(
  color = "#404040",
  family = "Avenir Next",
  fontface = "plain",
  size = 5,
  ...
)
}
\arguments{
\item{color}{An RGB hex string indicating the color to use for the text. The
default is "#404040".}

\item{family}{A string indicating the font family to use for the text. The
default is "Open Sans".}

\item{fontface}{A string indicating the font face to use for the text. The
default is "plain".}

\item{size}{An integer indicating the size of the text. The default is 5.}

\item{...}{Any other arguments passed to \code{annotate}.}
}
\description{
This function is identical to \code{\link[ggplot2]{annotate}}  in ggplot2.
The only difference is that the default arguments implement the pilot chart
style.
}
