% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{geom_text_pilot}
\alias{geom_text_pilot}
\title{A wrapper for the text geometry in ggplot2}
\usage{
geom_text_pilot(
  color = "#ffffff",
  family = "Avenir Next Demi Bold",
  fontface = "plain",
  ...
)
}
\arguments{
\item{color}{An RGB hex string indicating the color to use for the text. The
default is "#ffffff".}

\item{family}{A string indicating the font family to use for the text. The
default is "Avenir Next".}

\item{fontface}{A string indicating the font face to use for the text. The
default is "bold".}

\item{...}{Any other arguments passed to \code{geom_text}.}
}
\description{
This function is identical to \code{\link[ggplot2]{geom_text}} in ggplot2.
The only difference is that the default arguments implement the pilot chart
style.
}
