% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_pilot_colors}
\alias{get_pilot_colors}
\title{Function to extract pilot colors as hex codes}
\usage{
get_pilot_colors(...)
}
\arguments{
\item{...}{Names of colors in \code{pilot_colors}.}
}
\description{
Function to extract pilot colors as hex codes
}
