% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_pilot_palette}
\alias{get_pilot_palette}
\title{Return a function which interpolates a pilot color palette}
\usage{
get_pilot_palette(palette = "five", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette in \code{pilot_palettes}.}

\item{reverse}{Boolean to indicate whether the palette should be reversed.}

\item{...}{Additional arguments to pass to \code{colorRampPalette}.}
}
\description{
Return a function which interpolates a pilot color palette
}
