% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupdata}
\alias{groupdata}
\title{Data for SUV and cognitive scores (ESKD vs controls)}
\format{
## `groupdata`
A data frame with 20 rows and 17 columns:
\describe{
  \item{id}{Id variables}
  \item{group}{Group indicator, 0:Control, 1:ESKD}
  \item{age}{Age}
  \item{sex}{Sex}
  \item{bmi}{Body Mass Index}
  \item{mean_suv}{Standardized Uptake Values from SPECT-CT}
  \item{blind_moca_uncorrected}{Montreal Congitive Assessment (MoCA)}
  \item{craft_verbatim}{}
  \item{craft_delay_verbatim}{}
  \item{number_span_forward}{}
  \item{number_span_backward}{}
  \item{fluency_f_words_correct}{}
  \item{oral_trail_part_a}{}
  \item{oral_trail_part_b}{}
  \item{fluency_animals}{}
  \item{fluency_vegetables}{}
  \item{verbal_naming_cues}{}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
groupdata
}
\description{
Data from a pilot study comparing ESKD patients to controls
on SUV and a set of cognitive variables
}
\keyword{datasets}
