% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILOT.R
\name{pilot}
\alias{pilot}
\title{PIecewise LInear Organic Tree}
\usage{
pilot(
  X,
  y,
  dfs = c(1, 2, 5, 5, 7, 5),
  min_sample_leaf = 5,
  min_sample_alpha = 5,
  min_sample_fit = 10,
  maxDepth = 20,
  maxModelDepth = 100,
  rel_tolerance = 1e-04
)
}
\arguments{
\item{X}{an \eqn{n \times d} set of training data. Should be a data frame.}

\item{y}{a vector of length \eqn{n} with the training response}

\item{dfs}{a vector of length 6 indicating the "degrees of freedom" given to each of the models con/lin/pcon/blin/plin/pconc. A negative value results in the algorithm not considering this time of model while building the tree.}

\item{min_sample_leaf}{a positive integer. Do not consider splits which result in less than \code{min_sample_leaf} in a leaf node.}

\item{min_sample_alpha}{a positive integer. Do not consider splits which result in less than \code{min_sample_alpha} in a leaf node.}

\item{min_sample_fit}{a positive integer. Stop splitting the tree in a given node once no more than \code{min_sample_fit} observations are left in that node.}

\item{maxDepth}{an integer specifying the maximum depth of the tree. The depth of a tree is incremented when pcon/blin/plin/pconc models are fit.}

\item{maxModelDepth}{an integer specifying the maximum model depth of the tree. Model depth is incremented when any model is fit, including lin models.}

\item{rel_tolerance}{a lin model is only fit when the relative RSS is decreased by \code{rel_tolerance}.}
}
\value{
an object of the \code{PILOT} class, i.e. a list with the following components:
\itemize{
  \item \code{modelmat}: A matrix describing the fitted linear model tree.
  \item \code{residuals}: the residuals of the PILOT model on the training data
  \item \code{parameters}: A list containing the parameters used for fitting the model
  \item \code{data_names}: A vector with the column names of the training data.  Used for out-of-sample prediction.
  \item \code{catInfo}: A list containing information on the categorical variables in the data. Used for out-of-sample prediction.
  \item \code{modelpointer}: A pointer to the C++ object from the C++ class PILOT. 
  \item \code{jsonString}: A string describing the fitted linear model tree
}
}
\description{
Fits a linear model tree using the PILOT algorithm. PILOT builds a tree in a greedy way, much like CART, but incorporates linear models resulting in a linear model tree.
}
\examples{
data <- iris
y <- as.vector(data[, 1])
X <- as.data.frame(data[, 2:4])
pilot.out <- pilot(X, y)
# plot residuals
plot(pilot.out$residuals)
# generate predictions in-sample
preds.out <- predict(pilot.out, newdata = X)
plot(preds.out, y); abline(0, 1)
# print model matrix
print(pilot.out)
# plot pilot tree with models printed in leaf nodes
plot(pilot.out)
# plot pilot tree with a plot of variable importance in each leaf node
plot(pilot.out, infoType = 1)
}
