% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pilot_color}
\alias{pilot_color}
\title{Get the hex code for a pilot color}
\usage{
pilot_color(color_name)
}
\arguments{
\item{color_name}{The name of the color.}
}
\description{
Get the hex code for a pilot color
}
