% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predadjusted.R
\name{predadjusted}
\alias{predadjusted}
\title{Title}
\usage{
predadjusted(
  data,
  variables,
  covariates,
  id,
  timevar,
  type = "group",
  cor = "pearson"
)
}
\arguments{
\item{data}{Data set}

\item{variables}{Variables of interest}

\item{covariates}{Covariates to adjust for}

\item{id}{Id variable}

\item{timevar}{Variable denoting time, only used for pre-post data}

\item{type}{Type of analysis, pre-post, or group comparison}

\item{cor}{Type of correlation, default is Pearson}
}
\description{
Title
}
