% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predboot.R
\name{predboot}
\alias{predboot}
\title{Title}
\usage{
predboot(weights, results, nullphi = 0.5, alpha = 0.05, sims = 5000)
}
\arguments{
\item{weights}{A vector of weights for each endpoint}

\item{results}{A vector of 1's and 0's for each endpoint (the prediction results)}

\item{nullphi}{The null hypothesized phi value}

\item{alpha}{The type I error rate}

\item{sims}{The number of bootstrap simulations}
}
\value{
A list of outputs from the bootstrap test.
}
\description{
Title
}
\examples{

weights <- c(0.25, 0.3, 0.6, 0.15, 0.76, 0.17, 0.23)
results <- c(1, 1, 1, 0, 0, 1, 1)
nullphi <- 0.50
alpha <- 0.05
sims <- 100
}
