% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaFFLE.R
\name{predict.RAFFLE}
\alias{predict.RAFFLE}
\title{Predict with a RAFFLE model}
\usage{
\method{predict}{RAFFLE}(object, newdata, maxDepth = NULL)
}
\arguments{
\item{object}{an object of the RAFFLE class}

\item{newdata}{a matrix or data frame with new data}

\item{maxDepth}{predict using all nodes of depth up to maxDepth. If NULL, predict using full tree.}
}
\description{
Predict with a RAFFLE model
}
\examples{
data <- iris
y <- as.vector(data[, 1])
X <- as.data.frame(data[, 2:4])
raffle.out <- raffle(X, y, maxDepth = 3)
# generate predictions in-sample
preds.out <- predict(raffle.out, newdata = X)
plot(preds.out, y)

}
