% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predplot.R
\name{predplot}
\alias{predplot}
\title{Title}
\usage{
predplot(predictionresults, weights, order = FALSE)
}
\arguments{
\item{predictionresults}{A list, typically the output from the predresults function.}

\item{weights}{An optional vector of weights for the endpoints}

\item{order}{Whether to order the endpoints from largest to smallest}
}
\value{
A plot
}
\description{
Title
}
\examples{
results <- c(1,0,1,1,0,1,1,0)
diffs <- c(0.2, -0.5, 0.1, 1.2, -1, 0.65, 0.3, 0)
variables <- c("A", "B", "C", "D", "E", "F", "G", "H")
fromresults <- list(results, diffs, variables)
predplot(fromresults)
}
