% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predresults.R
\name{predresults}
\alias{predresults}
\title{Title}
\usage{
predresults(
  dataset,
  direction,
  bound = "wilcoxon",
  variables,
  type = "group",
  gtvar,
  phi_0 = 0.5,
  predictions,
  location = "mean"
)
}
\arguments{
\item{dataset}{Data to be used}

\item{direction}{Direction of prediction across endpoints. Increase (all increase), decrease (all decrease), mixed (provide a vector of predictions for each variable)}

\item{bound}{Whether or not a bound will be used}

\item{variables}{Endpoints of interest}

\item{type}{Type of analysis, pre-post or group}

\item{gtvar}{Variable denoting either the group or the time, dependent on type}

\item{phi_0}{The null hypothesized value}

\item{predictions}{A matrix with two columns. First column provides the variable names, second column the directional prediction.}

\item{location}{Measure of central tendency, mean or median}
}
\value{
A list of two elements. The first element is an indicator for whether each prediction on a variable was correct, the second element is the observed difference between groups or pre-post.
}
\description{
Title
}
