% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predtest.R
\name{predtest}
\alias{predtest}
\title{Title}
\usage{
predtest(weights, results, nullphi = 0.5, alpha = 0.05, exact = TRUE)
}
\arguments{
\item{weights}{A vector of weights for each endpoint}

\item{results}{A vector of 1's and 0's (the results for each endpoint)}

\item{nullphi}{Null hypothesized phi value}

\item{alpha}{Type I error rate}

\item{exact}{Should the calculation be exact? Supported for m < 20}
}
\value{
A printed list
}
\description{
Title
}
\examples{
weights <- c(0.25, 0.3, 0.6, 0.15, 0.76, 0.17, 0.23)
results <- c(1, 1, 1, 0, 0, 1, 1)
nullphi <- 0.50
alpha <- 0.05
exact <- TRUE
}
