% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILOT.R
\name{print.PILOT}
\alias{print.PILOT}
\title{Print a PILOT model}
\usage{
\method{print}{PILOT}(x, ...)
}
\arguments{
\item{x}{an object of the PILOT class}

\item{...}{other print parameters}
}
\description{
Print a PILOT model
}
\examples{
data <- iris
y <- as.vector(data[, 1])
X <- as.data.frame(data[, 2:4])
pilot.out <- pilot(X, y)
# print model matrix
print(pilot.out)

}
