% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaFFLE.R
\name{print.RAFFLE}
\alias{print.RAFFLE}
\title{Print a RAFFLE model}
\usage{
\method{print}{RAFFLE}(x, treeNb = 1, ...)
}
\arguments{
\item{x}{an object of the RAFFLE class}

\item{treeNb}{the tree that is printed. defaults to the first tree}

\item{...}{other print parameters}
}
\description{
Print a given tree in the RAFFLE model
}
\examples{
data <- iris
y <- as.vector(data[, 1])
X <- as.data.frame(data[, 2:4])
raffle.out <- raffle(X, y, maxDepth = 3)
 print(raffle.out, treeNb = 1)
 print(raffle.out, treeNb = 2)
}
