% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{save_png}
\alias{save_png}
\title{Save a plot as a png}
\usage{
save_png(plot, filename, width, height, dpi = 400)
}
\arguments{
\item{plot}{A plot object.}

\item{filename}{Filename to save to a png.}

\item{width}{Width of the image in inches.}

\item{height}{Height of the image in inches.}

\item{dpi}{Number of pixels per inch for the png. The default is 400.}
}
\description{
Save a plot as a png
}
