% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_pilot}
\alias{scale_color_pilot}
\title{Color scale for pilot colors}
\usage{
scale_color_pilot(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette in \code{pilot_palettes}.}

\item{discrete}{Boolean to indicate if color aesthetic is discrete.}

\item{reverse}{Boolean to indicate whether palette should be reversed.}

\item{...}{Additional arguments passed to \code{discrete_scale} or
\code{scale_color_gradientn}, depending on the value of \code{discrete}.}
}
\description{
Color scale for pilot colors
}
