% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_pilot}
\alias{scale_fill_pilot}
\title{Fill scale for pilot colors}
\usage{
scale_fill_pilot(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Name of palette in \code{pilot_palettes}.}

\item{discrete}{Boolean to indicate if color aesthetic is discrete.}

\item{reverse}{Boolean to indicate whether palette should be reversed.}

\item{...}{Additional arguments passed to \code{discrete_scale} or
\code{scale_color_gradientn}, depending on the value of \code{discrete}.}
}
\description{
Fill scale for pilot colors
}
