% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_pilot_family}
\alias{set_pilot_family}
\title{Set the font family used for all supported text elements}
\usage{
set_pilot_family(family)
}
\arguments{
\item{family}{The name of the font family to use.}
}
\description{
The font family used for each text element is controlled through a set of
options. This function sets each of these options to the same family,
allowing you to quickly customise the font family used by the theme.
Individual family options can be set with \code{options}.
}
