\name{pilot-package}
\alias{pilot-package}
\docType{package}
\title{
\packageTitle{pilot}
}
\description{
\packageDescription{pilot}
}
\details{
This package provides an implementation of the PILOT algorithm described in Raymaekers, Rousseeuw, Verdonck and Yao (2024) <doi:10.1007/s10994-024-06590-3> (open access).
}
\author{
\packageAuthor{pilot}

Maintainer: \packageMaintainer{pilot}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = rpilotdevel)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{rpilotdevel}}
%\section{Documentation Index}{\packageIndices{rpilotdevel}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
