% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILOT.R
\name{plot.PILOT}
\alias{plot.PILOT}
\title{Plot a PILOT model}
\usage{
\method{plot}{PILOT}(x, ...)
}
\arguments{
\item{x}{an object of the PILOT class}

\item{...}{other graphical parameters}
}
\description{
Plot a PILOT model
}
\examples{
x <- rnorm(10)

}
