% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaFFLE.R
\name{plot.RAFFLE}
\alias{plot.RAFFLE}
\title{Plot a RAFFLE model}
\usage{
\method{plot}{RAFFLE}(x, ...)
}
\arguments{
\item{x}{an object of the RAFFLE class}

\item{...}{other graphical parameters}
}
\description{
Plot a RAFFLE model
}
\examples{
x <- rnorm(10)

}
