% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predadjusted.R
\name{predadjusted}
\alias{predadjusted}
\title{Title}
\usage{
predadjusted(
  dataset,
  variables,
  covariates,
  id,
  type = "group",
  gtvar,
  phi_0 = 0.5,
  direction,
  predictions,
  corM = "sigma",
  location = "median"
)
}
\arguments{
\item{variables}{Variables of interest}

\item{covariates}{Covariates}

\item{id}{Id variable}

\item{type}{Type of analysis, pre-post, or group comparison}

\item{data}{Data set}

\item{timevar}{Variable denoting time, only used for pre-post data}

\item{cor}{Method to estimate correlation, sigma uses the estimated covariance matrix from multivariate regression for group and for pre-post the model estimated predictions are subtracted from the raw values and the correlation matrix is calculated from those, resid, uses the residuals (only for group comparisons).  .}
}
\value{
A list of vectors for results, observed differences, variable names and weights.
}
\description{
Title
}
