% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILOT.R
\name{predict.PILOT}
\alias{predict.PILOT}
\title{Plot a PILOT model}
\usage{
\method{predict}{PILOT}(object, newdata)
}
\arguments{
\item{object}{an object of the PILOT class}

\item{newdata}{a matrix or data frame with new data}
}
\description{
Plot a PILOT model
}
\examples{
x <- rnorm(10)

}
