% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaFFLE.R
\name{predict.RAFFLE}
\alias{predict.RAFFLE}
\title{Predict with a RAFFLE model}
\usage{
\method{predict}{RAFFLE}(x, newdata, maxDepth = NULL)
}
\arguments{
\item{x}{an object of the RAFFLE class}

\item{newdata}{a matrix or data frame with new data}

\item{maxDepth}{predict using all nodes of depth up to maxDepth. If NULL, predict using full tree.}
}
\description{
Predict with a RAFFLE model
}
\examples{
x <- rnorm(10)

}
