% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predplot.R
\name{predplot}
\alias{predplot}
\title{Title}
\usage{
predplot(predictionresults, weights, order = FALSE)
}
\arguments{
\item{predictionresults}{Output from the results function.}
}
\value{
A plot
}
\description{
Title
}
