% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predresults.R
\name{predresults}
\alias{predresults}
\title{Title}
\usage{
predresults(
  dataset,
  direction,
  bound = "wilcoxon",
  variables,
  type = "group",
  gtvar,
  phi_0 = 0.5,
  predictions,
  location = "mean"
)
}
\arguments{
\item{location}{}
}
\description{
Title
}
