% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predweights.R
\name{predweights}
\alias{predweights}
\title{Title}
\usage{
predweights(data, variables, id, timevar, type = "group", cor = "pearson")
}
\arguments{
\item{data}{Data set}

\item{variables}{Variables in the data set you want to include (only endpoints)}

\item{id}{The ID variable from the data set}

\item{timevar}{The time variable from the data set}

\item{type}{The type of analysis (pre-post or between groups)}

\item{cor}{Type of correlation, e.g. pearson}
}
\value{
A list of weights
}
\description{
Title
}
