% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predweights.R
\name{predweights}
\alias{predweights}
\title{Title}
\usage{
predweights(dataset, variables, id, type = "group", gtvar, cor = "pearson")
}
\arguments{
\item{dataset}{Data set}

\item{variables}{Variables in the data set you want to include (only endpoints)}

\item{id}{The ID variable from the data set}

\item{type}{The type of analysis (pre-post or between groups)}

\item{gtvar}{The group or time variable from the data set}

\item{cor}{Type of correlation, default is pearson}
}
\value{
A list of weights, named the original variable names (type: group), or diff.variable name (type: prepost)
}
\description{
Title
}
\examples{
ID <- 1:30
buildtesting <- as.data.frame(ID)
buildtesting$group <- c(rep(0,15), rep(1,15))
buildtesting$v1 <- rnorm(30, 0, 1)
buildtesting$v2 <- rnorm(30, 0, 1)
buildtesting$v3 <- rnorm(30, 0, 1)
buildtesting$v4 <- rnorm(30, 0, 1)

predweights(dataset= buildtesting, variables = c("v1", "v2", "v3", "v4"),
 id = "ID2", type = "group", gtvar = "group")

ID2 <- c(rep(1:15,2))
buildtesting2 <- as.data.frame(ID2)
buildtesting2$time2 <- c(rep(0,15), rep(12,15))
buildtesting2$v1 <- rnorm(30, 0, 1)
buildtesting2$v2 <- rnorm(30, 0, 1)
buildtesting2$v3 <- rnorm(30, 0, 1)
buildtesting2$v4 <- rnorm(30, 0, 1)

predweights(dataset= buildtesting2, variables = c("v1", "v2", "v3", "v4"),
 id = "ID2", type = "prepost", gtvar = "time2")

}
