% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILOT.R
\name{print.PILOT}
\alias{print.PILOT}
\title{Print a PILOT model}
\usage{
\method{print}{PILOT}(object)
}
\arguments{
\item{object}{an object of the PILOT class}
}
\description{
Print a PILOT model
}
\examples{
x <- rnorm(10)

}
