% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaFFLE.R
\name{print.RAFFLE}
\alias{print.RAFFLE}
\title{Print a RAFFLE model}
\usage{
\method{print}{RAFFLE}(x, ...)
}
\arguments{
\item{x}{an object of the RAFFLE class}

\item{...}{other print parameters}
}
\description{
Print a RAFFLE model
}
\examples{
x <- rnorm(10)

}
